test: openapi:validate
